+++ lib/ts/recipe/thirdparty/index.ts
@@ -24,8 +24,11 @@ import { getNormalisedUserContext } from "../../utils";
 
 import { RecipeComponentsOverrideContextProvider } from "./componentOverrideContext";
 import Apple from "./providers/apple";
+import Bitbucket from "./providers/bitbucket";
+import Discord from "./providers/discord";
 import Facebook from "./providers/facebook";
 import Github from "./providers/github";
+import Gitlab from "./providers/gitlab";
 import Google from "./providers/google";
 import ThirdParty from "./recipe";
 import { UserInput, GetRedirectionURLContext, PreAPIHookContext, OnHandleEventContext } from "./types";
@@ -170,10 +173,13 @@ export default class Wrapper {
     /*
      * Providers
      */
-    static Google = Google;
     static Apple = Apple;
-    static Facebook = Facebook;
+    static Bitbucket = Bitbucket;
+    static Discord = Discord;
     static Github = Github;
+    static Gitlab = Gitlab;
+    static Google = Google;
+    static Facebook = Facebook;
     static ComponentsOverrideProvider = RecipeComponentsOverrideContextProvider;
 }
 
@@ -195,9 +201,12 @@ const ThirdpartyComponentsOverrideProvider = Wrapper.ComponentsOverrideProvider;
 export {
     init,
     Apple,
+    Bitbucket,
+    Discord,
+    Github,
+    Gitlab,
     Google,
     Facebook,
-    Github,
     getStateAndOtherInfoFromStorage,
     setStateAndOtherInfoToStorage,
     getAuthorisationURLWithQueryParamsAndSetState,